% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netfacs-helpers.R
\name{add_inactive_single_units}
\alias{add_inactive_single_units}
\title{Add inactive (missing) single units}
\usage{
add_inactive_single_units(d, single.units)
}
\arguments{
\item{d}{A dataframe, result of \code{\link{probability_of_combination}}}

\item{single.units}{A character vector of single AUs}
}
\description{
Add inactive (missing) single units
}
