% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arules_explorer.R
\name{arules.explorer}
\alias{arules.explorer}
\title{Use arulesViz::ruleExplorer on NetFACS data}
\usage{
arules.explorer(netfacs.data)
}
\arguments{
\item{netfacs.data}{netfacs data list}
}
\value{
Starts 'shiny' application of arulesViz::ruleExplorer() function)
}
\description{
Applies 'shiny' application of arulesViz() package, which allows the user to set the support (probability of elements co-occurring), confidence (likelihood of B occurring when A is present), and lift (increase in likelihood of B if A is known), and extract association rules. Includes fancy plots
}
\examples{
data(netfacstestdata)
angry.face = netfacs(data = emotions_set[[1]],
 condition = emotions_set[[2]]$emotion,
 test.condition = 'anger',
 null.condition = NULL,
 duration = NULL, 
 ran.trials = 100,
 control = NULL,
 random.level = NULL,
 combination.size = 5)
arules.explorer(netfacs.data = netfacstestdata)
}
