% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arules_return.R
\name{arules.return}
\alias{arules.return}
\title{Prune and plot apriori ruleset}
\usage{
arules.return(netfacs.data, confidence = 0.8, support = 0.05)
}
\arguments{
\item{netfacs.data}{netfacs data list}

\item{confidence}{probability that, in the data, the second element occurs if the first element or combination is already in place}

\item{support}{probability that, in the data, the combination occurs}
}
\value{
Returns the pruned ruleset and plots it using the arulesViz package
}
\description{
Returns rules that satisfy specified 'support' (probability that A and B happen together) and 'confidence' (probability that B occurs if A is present) and plots them
Rules work by selecting the lowest-order sufficient explanation for the observed data. So, for example, if the combination of A and B explains the occurrence of C, but A alone also explains C, than the rule is A -> C. If A alone does not explain C, and B alone neither, then AB -> C
}
\examples{
data(emotions_set)
angry.face = netfacs(data = emotions_set[[1]],
 condition = emotions_set[[2]]$emotion,
 test.condition = 'anger',
 null.condition = NULL,
 duration = NULL, 
 ran.trials = 100,
 control = NULL,
 random.level = NULL,
 combination.size = 5)
arules.return(netfacs.data = angry.face, confidence = 0.8, support = 0.1)
}
