% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netfacs_helpers.R
\name{calculate_combination_size}
\alias{calculate_combination_size}
\title{Calculate combination size}
\usage{
calculate_combination_size(x)
}
\arguments{
\item{x}{A character vector of AU combinations, sep by _}
}
\value{

}
\description{
Calculate combination size
}
