% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_prob_of_comb.R
\name{calculate_prob_of_comb}
\alias{calculate_prob_of_comb}
\title{Calculate probabilities of single elements and combinations occurring}
\usage{
calculate_prob_of_comb(elements, maxlen)
}
\arguments{
\item{elements}{list with vectors for all elements observed together at each event}

\item{maxlen}{maximum size of combinations to be considered}
}
\value{
Function returns a dataframe with observed probabilities for each combination in the dataset
}
\description{
Calculate probabilities of single elements and combinations occurring
}
