% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_columns.R
\name{combine.columns}
\alias{combine.columns}
\title{Combine columns into one}
\usage{
combine.columns(data, output)
}
\arguments{
\item{data}{dataframe, containing one row per frame, one column per communication unit}

\item{output}{character vector with combinations of columns that should be turned into one column, separated by underscore. Should have the form 'AU1_AU2'}
}
\value{
Function returns the dataframe enter as 'data', but with specified columns aggregated and the original columns removed
}
\description{
Take frame-by-frame data and combine information of two columns, e.g. because they represent the same signal or because they overlap so much that they should be treated as one
}
\examples{
data(netfacstestdata)
head(netfacstestdata)
net.set = combine.columns(data = netfacstestdata, output = '25_26')
head(net.set)
 
}
