% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netfacs-helpers-na.R
\name{count_complete_cases}
\alias{count_complete_cases}
\title{Count complete cases}
\usage{
count_complete_cases(combination, m)
}
\arguments{
\item{combination}{A character vector of AU combinations, separated by "_"}

\item{m}{A binary matrix}
}
\value{
A numeric vector
}
\description{
If there are NAs in the data, the following function re-calculates the number
of possible events that combinations could have occurred in
}
