% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rule_set3.R
\name{create_rule_set3}
\alias{create_rule_set3}
\title{Take vector of elements and calculate probabilities of elements and combinations occurring}
\usage{
create_rule_set3(elements, maxlen)
}
\arguments{
\item{elements}{list with vectors for all elements observed together at each event}

\item{maxlen}{maximum size of combinations to be considered}
}
\value{
Function returns a dataframe with observed probabilities for each combination in the dataset
}
\description{
Underlying code is C++ based
}
\author{
Alex Mielke
}
