% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_contexts.R
\name{define_contexts}
\alias{define_contexts}
\title{Define truth for AUs active in different contexts}
\usage{
define_contexts(aus, n_active_aus, contexts = NULL, au_fidelity = 1)
}
\arguments{
\item{aus}{A character vector of AUs}

\item{n_active_aus}{A numeric vector, the same length as contexts, indicating
the number of AUs active per context.}

\item{contexts}{A character vector of contexts}

\item{au_fidelity}{A number between 1 and 0.5, indicating the probability
that an AU is active in a context.}
}
\value{
A matrix of probabilities with contexts in rows and AUs in columns
}
\description{
Define truth for AUs active in different contexts
}
