% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_joint_prob.R
\name{define_joint_prob}
\alias{define_joint_prob}
\title{Joint probability distribution of AUs}
\usage{
define_joint_prob(aus, n_jp = 2, min_jp = 0.5)
}
\arguments{
\item{aus}{A character vector of AUs}

\item{n_jp}{Number of joint probabilities >0}

\item{min_jp}{Minimum joint probability. Must be between 0 and 1}
}
\description{
Joint probability distribution of AUs
}
