% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specificity_increase.R
\name{element.specificity}
\alias{element.specificity}
\title{(Defunct) Tests how much each element increases the specificity of all
combinations it is in}
\usage{
element.specificity(netfacs.data)
}
\arguments{
\item{netfacs.data}{object resulting from \code{\link{netfacs}} function}
}
\description{
This function is defunct Please see \code{\link{specificity_increase}}
instead
}
