% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emotions_set-data.R
\docType{data}
\name{emotions_set}
\alias{emotions_set}
\title{Letter Data}
\format{An object of class.}
\usage{
data(emotions_set)
}
\description{
Data from the Cohn-Kanade database, FACS data and emotions for posed images
}
\examples{
data(emotions_set)
}
\references{
XXX
}
\keyword{datasets}
