% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy_overall.R
\name{entropy.overall}
\alias{entropy.overall}
\title{Compares the observed and expected information content of the dataset}
\usage{
entropy.overall(netfacs.data)
}
\arguments{
\item{netfacs.data}{object resulting from netfacs() function}
}
\value{
Function returns the ratio of observed entropy/ expected entropy. Expected entropy is based on randomization and represents the maximum entropy a dataset with the same properties as this one can reach. The further the ratio is from 1, the more ordered the data are.
}
\description{
Establishes how'ordered' the data is: values close to 0 indicate that combinations are highly repetitive and predictable, while values close to 1 indicate that combinations are equiprobable and prediction of future combinations is difficult
}
\examples{
### how do angry facial expressions differ from non-angry ones?
data(emotions_set)
angry.face = netfacs(data = emotions_set[[1]],
 condition = emotions_set[[2]]$emotion,
 test.condition = 'anger',
 null.condition = NULL,
 duration = NULL, 
 ran.trials = 100,
 control = NULL,
 random.level = NULL,
 combination.size = 5)
 
 entropy.overall(angry.face)
}
