% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy_slope.R
\name{entropy.slope}
\alias{entropy.slope}
\title{Creates the entropy on different levels}
\usage{
entropy.slope(netfacs.data, entropy.level = 5, random = F)
}
\arguments{
\item{netfacs.data}{object resulting from netfacs() function}

\item{entropy.level}{numeric value, indicating the maximum order up to which entropy should be established}

\item{random}{if F, the entropy slope for the data is produced; if T, the mean entropy slope of a randomisation of the dataset is also shown for comparison}
}
\value{
Function returns dataframe with different orders and entropy level on this order. Entropy should decrease with increasing order, as fewer degrees of freedom exist (i.e., if only one element is known, predictability for the next element is low; if three elements are known, the next element is usually quite predictable). Expected random entropy is the average entropy of 10 datasets with the same number of events and keeping the number of elements per event the same as in the observed data.
}
\description{
Establishes the n-th order entropies for the netfacs object. 0-Level entropy is the overall possible information content if all elements were equally distributed; first-order entropy is the possible information content gives the probabilities of each element; second-order entropy is the information content given the one element of a combination is known; etc. Decreasing entropy indicates higher predictability/fewer degrees of freedom.
}
\examples{
### how do angry facial expressions differ from non-angry ones?
data(emotions_set)
angry.face = netfacs(data = emotions_set[[1]],
 condition = emotions_set[[2]]$emotion,
 test.condition = 'anger',
 null.condition = NULL,
 duration = NULL, 
 ran.trials = 100,
 control = NULL,
 random.level = NULL,
 combination.size = 5)
 
 entropy.slope(angry.face, entropy.level = 5)
}
