% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netfacs-helpers.R
\name{equal_observations}
\alias{equal_observations}
\title{Check that ALL objects have the same number of observations}
\usage{
equal_observations(x, ...)
}
\arguments{
\item{x}{Object to compare number of observations}

\item{...}{Additional objects to compare number of observations}
}
\value{
Logical
}
\description{
lenght(vector), nrow(matrix), nrow(dataframe)
}
