% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netfacs-helpers.R
\name{get_active_elements}
\alias{get_active_elements}
\title{Extract active elements from matrix}
\usage{
get_active_elements(m)
}
\arguments{
\item{m}{A binary matrix where 1 indicates an element was active. colnames(m)
must contain the element names}
}
\value{
A list of vectors
}
\description{
Extract active elements from matrix
}
