% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netfacs-helpers-na.R
\name{get_active_elements_na}
\alias{get_active_elements_na}
\title{Extract active elements from matrix}
\usage{
get_active_elements_na(m)
}
\arguments{
\item{m}{A binary matrix where 1 indicates an element was active. colnames(m)
must contain the element names}
}
\value{
A list of vectors
}
\description{
Removes NA from result
}
