% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netfacs-class.R
\name{get_data}
\alias{get_data}
\title{Extract used data from a \code{netfacs} object}
\usage{
get_data(x, condition = "all")
}
\arguments{
\item{x}{extract data from the test condition of a \code{netfacs} object}

\item{condition}{one of "all" (default), "test" or "null".}
}
\description{
Extract used data from a \code{netfacs} object
}
