% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netfacs-class.R
\name{is.netfacs}
\alias{is.netfacs}
\title{Checks if argument is a \code{netfacs} object}
\usage{
is.netfacs(x)
}
\arguments{
\item{x}{An \R object}
}
\description{
Checks if argument is a \code{netfacs} object
}
