% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netfacs-class.R
\name{is.netfacs_multiple}
\alias{is.netfacs_multiple}
\title{Checks if argument is a \code{netfacs_multiple} object}
\usage{
is.netfacs_multiple(x)
}
\arguments{
\item{x}{An \R object}
}
\description{
Checks if argument is a \code{netfacs_multiple} object
}
