% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netfacs-class.R
\name{is.netfacs_specificity}
\alias{is.netfacs_specificity}
\title{Checks if argument is a \code{netfacs_specificity} object}
\usage{
is.netfacs_specificity(x)
}
\arguments{
\item{x}{An \R object}
}
\description{
Checks if argument is a \code{netfacs_specificity} object
}
