% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/letternet-data.R
\docType{data}
\name{letternet}
\alias{letternet}
\title{Letter Data}
\format{
An object of class.
}
\usage{
data(letternet)
}
\description{
Data from the German, English, and French Versions of The Communist Manifesto, to have large datasets to test different functions in this package for now
}
\references{
Marx & Engels, 'The Communist Manifesto'
}
\keyword{datasets}
