% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netfacs_clustering.R
\name{netfacs.clustering}
\alias{netfacs.clustering}
\title{Dimension reduction and unsupervised learning of conditions}
\usage{
netfacs.clustering(netfacs.data, condition = NULL)
}
\arguments{
\item{netfacs.data}{object resulting from netfacs() function}

\item{condition}{vector containing information about the classification of known cases. If NULL, unsupervised learning is produced}
}
\value{
Returns a UMAP object that can be used to predict where new cases cluster. Also produces graph of unsupervised clustering, table of overlap of condition and unsupervised clusters
}
\description{
Takes full dataset from netfacs function, conducts UMAP (uniform manifold approximation and projection) to reduce dimensions and then k-Means to detect optimal number of categories. Can also test those categories against used categories and can be used to add new cases to see where they fall.
}
\examples{
angry.face = netfacs(data = emotions_set[[1]],
 condition = emotions_set[[2]]$emotion,
 test.condition = 'anger',
 null.condition = NULL,
 duration = NULL, 
 ran.trials = 100,
 control = NULL,
 random.level = NULL,
 combination.size = 5)
 
 angry.umap = netfacs.clustering(angry.face, condition = angry.face$used.data$condition)
 angry.umap$plot
 angry.umap$comparison.table
}
