% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netfacs_extract.R
\name{netfacs.extract}
\alias{netfacs.extract}
\title{(Deprecated) Extract results from a \code{\link{netfacs}} object.}
\usage{
netfacs.extract(
  netfacs.data,
  combination.size = NULL,
  significance = 1,
  min.count = 0,
  min.prob = 0,
  min.specificity = 0,
  level
)
}
\arguments{
\item{netfacs.data}{An object of class \code{\link{netfacs}}.}

\item{combination.size}{Numeric, denoting the combination size(s) that should
be extracted. If \code{NULL} (default), all combination sizes are returned.}

\item{significance}{Numeric value between 0 and 1, determining the p-value
below which combinations are considered to be dissimilar enough from the
null distribution.}

\item{min.count}{Numeric, denoting the minimum number of times an element
combination occurred.}

\item{min.prob}{Numeric value between 0 and 1, denoting the minimum
probability an element combination occurred to be displayed.}

\item{min.specificity}{Numeric value between 0 and 1, denoting the minimum
specificity an element combination and the test condition.}

\item{level}{deprecated. Please use combination.size instead.}
}
\value{
Function returns a dataframe that contains the results of the
  \code{\link{netfacs}} object. By default, returns all results for all
  observed combinations
}
\description{
This function is deprecated. Please see \code{\link{netfacs_extract}}
instead
}
