% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netfacs_tsne.R
\name{netfacs.tsne}
\alias{netfacs.tsne}
\title{Conducts dimension reduction using t-distributed stochastic neighbor embedding and plots conditions}
\usage{
netfacs.tsne(netfacs.data)
}
\arguments{
\item{netfacs.data}{object resulting from netfacs() function}
}
\value{
Function returns a ggnet plot where each condition is connected to those elements that occur significantly in this condition, and each element is connected to each condition under which it occurs significantly more than expected
}
\description{
Takes full dataset from netfacs function, conducts tsne using Rtsne package, and plots results
}
\examples{
angry.face = netfacs(data = emotions_set[[1]],
 condition = emotions_set[[2]]$emotion,
 test.condition = 'anger',
 null.condition = NULL,
 duration = NULL, 
 ran.trials = 100,
 control = NULL,
 random.level = NULL,
 combination.size = 5)
 
 entropy.slope(angry.face, entropy.level = 5)
}
