% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netfacs-helpers.R
\name{netfacs_randomize}
\alias{netfacs_randomize}
\title{Calculate probabilities from single randomization}
\usage{
netfacs_randomize(m, test.combinations, max.combination.size, max.event.size)
}
\arguments{
\item{m}{A numeric matrix}

\item{test.combinations}{A vector of AU combinations observed in test data}

\item{max.combination.size}{A positive integer}

\item{max.event.size}{A Positive integer}
}
\value{
A list of randomized probabilities for combinations and event sizes
}
\description{
Calculate probabilities from single randomization
}
