% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_plot.R
\name{network.plot}
\alias{network.plot}
\title{Plots a network object}
\usage{
network.plot(netfacs.graph, title = "network", clusters = T)
}
\arguments{
\item{netfacs.graph}{igraph network object resulting from netfacs.network() function}

\item{title}{string of the graph's main title}

\item{clusters}{if TRUE, the igraph::cluster_fast_greedy function is used to establish possible clusters in the dataset}
}
\value{
Function returns a ggnet plot of the network, where the size of nodes indicates how often they occur on their own, and edges indicate significant co-occurrance between them
}
\description{
Plots the network created using the netfacs.network() function; for networks with clear clusterin of elements, clusters can get different colours
}
\examples{
data(emotions_set)
angry.face = netfacs(data = emotions_set[[1]],
 condition = emotions_set[[2]]$emotion,
 test.condition = 'anger',
 null.condition = NULL,
 duration = NULL, 
 ran.trials = 100,
 control = NULL,
 random.level = NULL,
 combination.size = 2)
 
 anger.net = netfacs.network(netfacs.data = angry.face, package = 'igraph', link = 'unweighted', significance = 0.01, min.count = 1, min.prob = 0, min.specificity = 0, ignore.element = NULL)
 anger.plot = network.plot(anger.net, title = 'Angry Faces', clusters = F)
 anger.plot + ggtitle('This is a ggplot object and can be manipulated')
 
 all.face = netfacs(data = emotions_set[[1]],
 ran.trials = 100,
 combination.size = 2)
 all.net = netfacs.network(netfacs.data = all.face, package = 'igraph', link = 'unweighted', significance = 0.01, min.count = 1, min.prob = 0, min.specificity = 0, ignore.element = NULL)
 all.plot = network.plot(anger.net, title = 'All Emotions', clusters = F)
}
