% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_plot.R
\name{network_plot}
\alias{network_plot}
\alias{network.plot}
\title{Plots a network object}
\usage{
network_plot(
  netfacs.graph,
  title = "network",
  clusters = FALSE,
  plot.bubbles = FALSE,
  hide.unconnected = TRUE
)

network.plot(
  netfacs.graph,
  title = "network",
  clusters = FALSE,
  plot.bubbles = FALSE,
  hide.unconnected = TRUE
)
}
\arguments{
\item{netfacs.graph}{igraph network object resulting from
\code{\link{netfacs_network}}}

\item{title}{string of the graph's main title}

\item{clusters}{if \code{TRUE}, \code{\link[igraph]{cluster_fast_greedy}} is
used to establish possible clusters in the dataset}

\item{plot.bubbles}{if TRUE, then the nodes in the network plots will be
surrounded by bubbles; if FALSE, the edges connect the names directly}

\item{hide.unconnected}{if TRUE, then the nodes that do not have any
significant connections will be hidden in the plot}
}
\value{
Function returns a \code{\link[ggraph]{ggraph}} plot of the network,
  where the size of nodes indicates how often they occur on their own, and
  edges indicate significant co-occurrence between them
}
\description{
Plots the network created using the \code{\link{netfacs_network}} function;
for networks with clear clusterin of elements, clusters can get different
colours
}
\examples{
data(emotions_set)
angry.face <- netfacs(
  data = emotions_set[[1]],
  condition = emotions_set[[2]]$emotion,
  test.condition = "anger",
  ran.trials = 10,
  combination.size = 2
)

anger.net <- netfacs_network(
  netfacs.data = angry.face,
  link = "unweighted",
  significance = 0.01,
  min.count = 1
)

network_plot(anger.net,
             title = "Angry Faces",
             clusters = FALSE,
             plot.bubbles = TRUE)
}
