% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_network.R
\name{overlap.network}
\alias{overlap.network}
\title{Plots the overlap of multiple conditions as bipartite network}
\usage{
overlap.network(
  netfacs.list,
  min.prob = 0,
  min.count = 5,
  significance = 0.01,
  ignore.element = NULL
)
}
\arguments{
\item{netfacs.list}{list of objects resulting from netfacs() function or multiple.netfacs() function}
}
\value{
Function returns a ggnet plot where each condition is connected to those elements that occur significantly in this condition, and each element is connected to each condition under which it occurs significantly more than expected
}
\description{
The function takes multiple netfacs objects and plots how different elements connect the conditions
}
\examples{
data(emotions_set)
emo.faces = multiple.netfacs(data = emotions_set[[1]],
 condition = emotions_set[[2]]$emotion,
 duration = NULL, 
 ran.trials = 100,
 control = NULL,
 random.level = NULL,
 combination.size = 5)
 
 overlap.network(emo.faces)
}
