% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netfacs-helpers.R
\name{possible_combinations}
\alias{possible_combinations}
\title{Calculate all possible combinations of elements}
\usage{
possible_combinations(elements, maxlen, sep = "_")
}
\arguments{
\item{elements}{A vector of elements}

\item{maxlen}{maximum size of combinations to be considered}

\item{sep}{String. Separator used for showing combinations of elements}
}
\value{
A vector with all element combinations
}
\description{
Takes a vector of elements and returns a vector with all possible
combinations
}
