% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.netfacs_multiple}
\alias{print.netfacs_multiple}
\title{Print method for objects of class \code{netfacs_multiple}}
\usage{
\method{print}{netfacs_multiple}(x, ...)
}
\arguments{
\item{x}{An object of class \code{netfacs_multiple}}

\item{...}{Additional arguments that would be passed to or from other methods}
}
\description{
Print method for objects of class \code{netfacs_multiple}
}
