% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netfacs-helpers-na.R
\name{probability_of_combination_na}
\alias{probability_of_combination_na}
\title{Calculate probabilities of single elements and combinations occurring}
\usage{
probability_of_combination_na(m, maxlen, sep = "_")
}
\arguments{
\item{m}{A binary matrix where 1 indicates an element was active. colnames(m)
must contain the element names}

\item{maxlen}{maximum size of combinations to be considered}

\item{sep}{String. Separator used for showing combinations of elements}
}
\value{
Function returns a dataframe with observed probabilities for each
  combination in the dataset
}
\description{
Calculate probabilities of single elements and combinations occurring
}
