% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netfacs-helpers.R
\name{probability_of_event_size}
\alias{probability_of_event_size}
\title{Count number of event sizes}
\usage{
probability_of_event_size(elements, max.event.size)
}
\arguments{
\item{elements}{A list of vectors containing active elements or a binary
matrix with events in rows}

\item{max.event.size}{A positive integer}
}
\value{
A named vector, including probabilities for event sizes that were not
  observed in the data
}
\description{
Count number of event sizes
}
