% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netfacs-helpers-na.R
\name{rand_replace_na}
\alias{rand_replace_na}
\title{Randomly assign 0s and 1s to NA}
\usage{
rand_replace_na(x)
}
\arguments{
\item{x}{A numeric vector}
}
\description{
Randomly assign 0s and 1s to NA
}
