% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_contexts.R
\name{sample_contexts}
\alias{sample_contexts}
\title{Sample observations of simulated FACS data}
\usage{
sample_contexts(m, n_obs = 10, jp = NULL)
}
\arguments{
\item{m}{A matrix with context in rows, aus in cols, and probabilities as
values => result of \code{\link{define_contexts}}.}

\item{n_obs}{Number of observations per context}

\item{jp}{An optional list of matrices, the same length as nrow(m) with the joint probabilities of AUs}
}
\value{

}
\description{
Sample observations of simulated FACS data
}
\examples{
m <- define_contexts(aus = as.character(1:5),
                     contexts = letters[1:5], 
                     n_active_aus = 1:5, 
                     au_fidelity = 1)

sample_contexts(m, n_obs = 10)
}
