% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specificity.R
\name{specificity}
\alias{specificity}
\title{Specificity}
\usage{
specificity(
  x,
  condition,
  test_condition = NULL,
  null_condition = NULL,
  balance_conditions = TRUE
)
}
\arguments{
\item{x}{A binary matrix, with AUs as \code{colnames()}}

\item{condition}{A character condition vector}

\item{test_condition}{A string, denoting the test condition}

\item{null_condition}{A string, denoting the null condition. If \code{NULL}
(default) all observations not part of the test_condition will be
considered part of the null.}
}
\value{
Tibble
}
\description{
Calculate specificity of element combinations to a given condition
}
\examples{
specificity(
  x = emotions_set[[1]],
  condition = emotions_set[[2]]$emotion,
  test_condition = "anger"
)
}
