% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netfacs-helpers-na.R
\name{split_combination}
\alias{split_combination}
\title{Split string separated by "_"}
\usage{
split_combination(x)
}
\arguments{
\item{x}{A string of AUs separated by "_"}
}
\value{
A character vector
}
\description{
Split string separated by "_"
}
