% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netfacs-helpers.R
\name{summarise_combination}
\alias{summarise_combination}
\title{Summarise combination results from bootstrap}
\usage{
summarise_combination(
  combination,
  combination.size,
  observed.prob,
  boot.prob,
  tail,
  test.count,
  null.count = NULL
)
}
\arguments{
\item{combination}{A vector of AU combinations}

\item{combination.size}{A vector denoting the number of active AUs in combination}

\item{observed.prob}{A vector with probability of combination in test data}

\item{boot.prob}{A matrix with boot probabilities of a given combination in
columns}

\item{tail}{upper.tail or lower.tail,}

\item{test.count}{Number of times a combination occurs in test dataset}

\item{null.count}{Number of times a combination occurs in null dataset}
}
\value{
A dataframe
}
\description{
Summarise combination results from bootstrap
}
