% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netfacs-helpers.R
\name{summarise_event_size}
\alias{summarise_event_size}
\title{Summarise event size probabilities}
\usage{
summarise_event_size(observed.prob, boot.prob)
}
\arguments{
\item{observed.prob}{A named vector with probabilities of event sizes.}

\item{boot.prob}{A matrix with boot probabilities of a given event
size. Combination size in rows, trials in columns.}
}
\value{
A dataframe
}
\description{
Summarise event size probabilities
}
