% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_pairs.R
\name{summarise_pairs}
\alias{summarise_pairs}
\title{Summarise dyadic combination of elements}
\usage{
summarise_pairs(netfacs.data)
}
\arguments{
\item{netfacs.data}{An object of class \code{\link{netfacs}}}
}
\value{
A summary data.frame
}
\description{
For all dyadic combinations that appear in the test dataset, this function
returns the probability of A occurring (P(A)), the probability of B occurring
(P(B)), the probability of A and B occurring simultaneously (P(AandB)), the
probability of A given B (P(A|B)), pointwise mutual information (pmi) and
normalized pointwise mutual information (norm_pmi). The normalized pmi ranges
between -1 and 1; -1 for never occurring together, 0 for independence, and +1
for complete co-occurrence.
}
\examples{
data(emotions_set)
angry.face <- netfacs(
  data = emotions_set[[1]],
  condition = emotions_set[[2]]$emotion,
  test.condition = "anger",
  ran.trials = 100,
  combination.size = 2
)

summarise_pairs(angry.face)
}
