% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netfacs-helpers.R
\name{validate_condition}
\alias{validate_condition}
\title{Check that condition arguments are formatted correctly}
\usage{
validate_condition(data, condition, test.condition, null.condition)
}
\arguments{
\item{data}{data passed by the user}

\item{condition}{condition passed by the user}

\item{test.condition}{condition passed by the user}

\item{null.condition}{condition passed by the user}
}
\description{
Check that condition arguments are formatted correctly
}
