% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netfacs-helpers.R
\name{validate_data}
\alias{validate_data}
\title{Check that 'data' argument is formatted correctly}
\usage{
validate_data(data)
}
\arguments{
\item{data}{data passed by the user}
}
\value{
data as a matrix
}
\description{
Check that 'data' argument is formatted correctly
}
