% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element_plot.R
\name{element.plot}
\alias{element.plot}
\title{Plots the probability increase for the basic elements based on the condition}
\usage{
element.plot(netfacs.data)
}
\arguments{
\item{netfacs.data}{object resulting from netfacs() function}
}
\value{
Function returns a ggplot showing for each element by how much the observed probability outperforms the expected probabilities
}
\description{
The function takes all single elements in a netfacs object, and plots the distribution of ratios between the observed value and all randomisations
}
\examples{
### how do angry facial expressions differ from non-angry ones?
data(emotions_set)
angry.face = netfacs(data = emotions_set[[1]],
 condition = emotions_set[[2]]$emotion,
 test.condition = 'anger',
 null.condition = NULL,
 duration = NULL, 
 ran.trials = 100,
 control = NULL,
 random.level = NULL,
 combination.size = 5)
 
 element.plot(angry.face)
}
