% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emotions_set-data.R
\docType{data}
\name{emotions_set}
\alias{emotions_set}
\title{Letter Data}
\format{
An object of class.
}
\usage{
data(emotions_set)
}
\description{
Data from the Extended Cohn-Kanade database, FACS data and emotions for posed images
}
\references{
Lucey P, Cohn JF, Kanade T, Saragih J, Ambadar Z, Matthews I (2010) The extended Cohn-Kanade dataset (CK+): A complete dataset for action unit and emotion-specified expression. In: 2010 IEEE Computer Society Conference on Computer Vision and Pattern Recognition - Workshops, CVPRW 2010. pp 94-101
}
\keyword{datasets}
