% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy_overall.R
\name{entropy.overall}
\alias{entropy.overall}
\title{(Deprecated) Calculate information content of the dataset}
\usage{
entropy.overall(x, netfacs.data)
}
\arguments{
\item{x}{An object of class \code{\link{netfacs}} or simply a binary matrix
of 0s and 1s, with elements in columns and events in rows.}

\item{netfacs.data}{deprecated. Please use x instead.}
}
\description{
This function is deprecated. Please see \code{\link{entropy_overall}}
instead
}
