% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netfacs_extract.R
\name{netfacs.extract}
\alias{netfacs.extract}
\title{(Deprecated) Extract results from a \code{\link{netfacs}} object.}
\usage{
netfacs.extract(
  netfacs.data,
  combination.size = NULL,
  significance = 1,
  min.count = 0,
  min.prob = 0,
  min.specificity = 0,
  level
)
}
\arguments{
\item{netfacs.data}{object resulting from \code{\link{netfacs}} function.}

\item{combination.size}{numeric, denoting the combination size(s) that should
be extracted. If NULL (default), all combination sizes are returned.}

\item{significance}{numeric value between 0 and 1, determining the p-value
below which combinations are considered to be dissimilar enough from the
null distribution.}

\item{min.count}{numeric value, suggesting how many times a combination
should at least occur to be displayed.}

\item{min.prob}{numeric value between 0 and 1, suggesting the probability at
which a combination should at least occur to be displayed.}

\item{min.specificity}{numeric value between 0 and 1, suggesting the
specificity a combination should at least have for the test condition to be
displayed.}

\item{level}{deprecated. Please use combination.size instead.}
}
\value{
Function returns a dataframe that contains the results of the
  \code{\link{netfacs}} object. By default, returns all results for all
  observed combinations
}
\description{
This function is deprecated. Please see \code{\link{netfacs_extract}}
instead
}
