% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_conditional.R
\name{network.conditional}
\alias{network.conditional}
\title{(Deprecated) Produce conditional probabilities of dyads of elements, and
graph object based on conditional probabilities}
\usage{
network.conditional(
  netfacs.data,
  min.prob = 0,
  min.count = 0,
  ignore.element = NULL,
  plot.bubbles = FALSE
)
}
\arguments{
\item{netfacs.data}{object resulting from \code{\link{netfacs}} or
\code{\link{conditional_probabilities}} functions.}

\item{min.prob}{minimum conditional probability that should be shown in the
graph}

\item{min.count}{minimum number of times that a combination should occur
before being included in the graph}

\item{ignore.element}{string vector, can be used to exclude certain elements
when creating the plots}

\item{plot.bubbles}{if TRUE (default), then the nodes in the network plots
will be surrounded by bubbles; if FALSE, the edges connect the names
directly}
}
\value{
Function returns a dataframe that includes all dyadic combinations
  and their observed and conditional probabilities
}
\description{
This function is deprecated. Please see \code{\link{network_conditional}}
instead
}
