% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_network.R
\name{overlap.network}
\alias{overlap.network}
\title{(Deprecated) Plots the overlap of multiple conditions as bipartite network.}
\usage{
overlap.network(
  netfacs.list,
  min.prob = 0,
  min.count = 5,
  significance = 0.01,
  specificity = 0.1,
  ignore.element = NULL,
  clusters = FALSE,
  plot.bubbles = FALSE
)
}
\arguments{
\item{netfacs.list}{list of objects resulting from \code{\link{netfacs}} or
\code{\link{netfacs_multiple}}}

\item{min.prob}{minimum conditional probability that should be shown in the
graph}

\item{min.count}{minimum number of times that a combination should occur
before being included in the graph}

\item{significance}{sets the level of significance that combinations have to
pass before added to the network}

\item{specificity}{for the 'reduced' graph, select only elements that surpass
this context specificity value}

\item{ignore.element}{string vector, can be used to exclude certain elements
when creating the plots}

\item{clusters}{boolean; if TRUE, the cluster_fast_greedy algorithm is used
to detect underlying community structure, based on the occurrence
probability network}

\item{plot.bubbles}{if TRUE, then the nodes in the network plots will be
surrounded by bubbles; if FALSE, the edges connect the names directly}
}
\description{
This function is deprecated. Please see \code{\link{overlap_network}}
instead
}
