% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netfacs-helpers.R
\name{probability_of_combination}
\alias{probability_of_combination}
\title{Calculate probabilities of single elements and combinations occurring}
\usage{
probability_of_combination(elements, maxlen)
}
\arguments{
\item{elements}{list with vectors for all elements observed together at each
event}

\item{maxlen}{maximum size of combinations to be considered}
}
\value{
Function returns a dataframe with observed probabilities for each
  combination in the dataset
}
\description{
Calculate probabilities of single elements and combinations occurring
}
