% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_contexts.R
\name{sample_contexts}
\alias{sample_contexts}
\title{Sample observations of simulated FACS data}
\usage{
sample_contexts(m, n_obs = 10, jp = NULL)
}
\arguments{
\item{m}{A matrix with context in rows, aus in cols, and probabilities as
values => result of \code{\link{define_contexts}}.}

\item{n_obs}{Number of observations per context}

\item{jp}{An optional list of matrices, the same length as nrow(m) with the joint probabilities of AUs}
}
\description{
Sample observations of simulated FACS data
}
